class7: add a top-level CMakeLists.txt
Kinda atypically, in the class repo, there's a lot of mostly independent
per-class directories, so it makes sense for them to have their own
independent build setup, etc.  But, an IDE like VS Code typically
looks for a top-level `CMakeLists.txt` file to help with building and
running the project.  So, let's add one.